% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroMeteorology-package.R
\docType{package}
\name{HydroMeteorology-package}
\alias{HydroMeteorology}
\alias{HydroMeteorology-package}
\title{HydroMeteorology: CUG-HydroMeteorology}
\description{
CUG-HydroMeteorology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CUG-hydro/CUG-HydroMeteorology}
  \item Report bugs at \url{https://github.com/CUG-hydro/CUG-HydroMeteorology/issues}
}

}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
